Project Description
The Disease Prediction and QA (Question-Answering) System in the Healthcare Domain is an ambitious and vital project aimed at leveraging the power of cutting-edge technology to address critical healthcare challenges. In the modern healthcare landscape, timely disease prediction and accurate information dissemination are of paramount importance. This project represents a convergence of data science, machine learning, natural language processing, and healthcare expertise to create a comprehensive solution.